Update linmix.py
This is an edited version of linmix that takes Gaussian mixture model fits to the individual object joint x-y posterior probability distribution functions.  It's a bit ugly at present.  It would be best to ensure the original use could still be made as now measured x- and y- values cannot be read in as single gaussians, measurements and posterior probabilities have to be treated differently.

TODO: allow for different use-cases; measurements plus errors or GMM fits to posterior probabilities